+++ smarthome_bridge/api_decoder.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from logging_interface import LoggingInterface
+from lib.logging_interface import LoggingInterface
 from gadgets.gadget import Gadget
 from system.gadget_definitions import GadgetIdentifier
 from smarthome_bridge.characteristic import Characteristic, CharacteristicIdentifier
